detect_set_scale: use ratio of drawable size and window size for auto-detection
On Fedora37, wayland with desktop scale set to 200% on a 4K screen
SDL2 would detect a DPI of 92.

https://wiki.libsdl.org/SDL2/SDL_GetDisplayDPI says that the function is
unreliable, and that the ratio of SDL_GL_GetDrawableSize and
SDL_GetWindowSize should be used instead.

There is already a dpi_xscale/dpi_yscale that is set, but setting DPI at
that point would be too late: the layout and fonts would already be
loaded with the old scale.

The only way to get the desired scale was to use BOGUE_SCALE=2 env var.
(Calling the various resize/set_size functions doesn't correctly refresh
everything with the new scale, apparently has to be done at
initialization time).

So create a fake 10 pixel window to detect scale for now.

For best cross platform compatibility use the ratio computed above as a
multiplier for the DPI obtained from the system.

This only works reliably with wayland: X11, Mac OS X and other platforms
already seem to report the correct DPI.

Signed-off-by: Edwin Török <edwin@etorok.net>