one_step: limit NOP event rate to 60 FPS
Sometimes SDL can receive a lot of events (e.g. 1000Hz mouse),
and we want to limit the frame-rate when not animating.

Previously the limit would've been ~10ms between Sdl.poll_event + 5ms
from Thread.delay ~66 FPS.

When the events have no effect on the display (nothing is rendered,
and in fact Sdl.render_present will be skipped too) limit the processing
rate to 60 Hz to match typical monitor refresh rate.

When we do receive some useful events that change the display then use
just the old 5ms Thread sleep as a throttling mechanism, but do it after
we've rendered the frame to reduce input-to-display latency.

Signed-off-by: Edwin Török <edwin@etorok.net>