get_dpi: use xrandr to get DPI on wayland
SDL2+wayland in various versions reports different DPI depending on
scaling.
On some versions it reports the monitor's true DPI regardless of desktop
scale factor. In other versions it scales the monitor's DPI down based
on desktop scale factor.
However there isn't a way to detect which one it is, so the only
reliable way to get DPI on wayland ... is to rely on XWayland and xrandr
which reports all this correctly.

(or use wayland-info but that isn't always installed).

'xdpyinfo' would report a bogus 96 DPI always, so that cannot be used.

Signed-off-by: Edwin Török <edwin@etorok.net>